requirements.txt
fastapi
uvicorn[standard]
python-dotenv
anthropic
pydantic
python-multipart